smart-school-project-main/smart_school_backend/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
atomicwrites==1.4.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.0
colorama==0.4.6
deepface==0.0.96
dlib==20.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
gast==0.7.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.3.0
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.15.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
libclang==18.1.1
lightecc==0.0.3
lightphe==0.0.19
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.2
py==1.11.0
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tomli==2.3.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
Werkzeug==3.1.4
wrapt==2.0.1
